migrate to long type

+++ commons/src/main/kotlin/com/simplemobiletools/commons/activities/AboutActivity.kt
@@ -299,7 +299,7 @@ class AboutActivity : BaseSimpleActivity() {
             Intent(applicationContext, LicenseActivity::class.java).apply {
                 putExtra(APP_ICON_IDS, getAppIconIDs())
                 putExtra(APP_LAUNCHER_NAME, getAppLauncherName())
-                putExtra(APP_LICENSES, intent.getIntExtra(APP_LICENSES, 0))
+                putExtra(APP_LICENSES, intent.getLongExtra(APP_LICENSES, 0))
                 startActivity(this)
             }
         }
